feat: dump local filename and checksums in addition to repository URLs
BREAKING CHANGE: this changes the output file format to YAML
This change of output format accommodates having more than a single
field per entry.

In addition to the destination URL we now provide the local file's path
and several checksums of its content. This extra meta data should
facilitate aggregating these artifacts' into "builds" in Artifactory.
This facilitates it in a way that doesn't depend on the absence of race
conditions between the time of upload through Ivy and the time that the
"build info" gets created.